// server/src/common/enums/error-code.enum.ts
export enum ErrorCode {
  UNEXPECTED_ERROR = 'UNEXPECTED_ERROR',
  INVALID_TOKEN = 'INVALID_TOKEN',
  TOKEN_EXPIRES = 'TOKEN_EXPIRES',
  NOT_FOUND_AUTH_USER = 'NOT_FOUND_AUTH_USER',
  NOT_FOUND_USER = 'NOT_FOUND_USER',
  NO_SERVICE = 'NO_SERVICE',
  USER_EXISTED = 'USER_EXISTED',
  INVALID_REQUEST = 'INVALID_REQUEST',
  INVALID_PASSWORD = 'INVALID_PASSWORD',
  NOT_FOUND_QUIZ = 'NOT_FOUND_QUIZ',
  INCORRECT_ANSWER = 'INCORRECT_ANSWER',
  NOT_FOUND_QUESTION = 'NOT_FOUND_QUESTION',
  UNAUTHORIZED_PLAY_QUIZ = 'UNAUTHORIZED_PLAY_QUIZ',
  USER_PLAYED_QUIZ = 'USER_PLAYED_QUIZ',
}